Create Using Two Queues implemented stack
Using
Two Queues
The idea is to implement the queue’s enqueue operation such that
the last entered item always ends up at the queue’s front. To achieve this, we
need an additional queue.
1.
To push an item into the stack, first move all elements from the
first queue to the second queue, then enqueue the new item into the first
queue, and finally move all elements back to the first queue. This ensures that
the new item lies in front of the queue and hence would be the first one to be
removed.
2.
To pop an item from the stack, return the front item from the
first queue.